test: 일최고, 최저온도 타입 int -> double

+++ backend/domain-modules/forecast-domain/src/test/java/com/softeer/repository/impl/ForecastJdbcRepositoryTest.java
@@ -200,7 +200,7 @@ class ForecastJdbcRepositoryTest {
     }
 
     private Long persistShortForecast(ForecastEntity forecastEntity, double precipitationProbability, double snowAccumulation,
-                                      int highestTemperature, int lowestTemperature
+                                      double highestTemperature, double lowestTemperature
                                       ) {
         ForecastEntity forecast = forecastJpaRepository.saveAndFlush(forecastEntity);
